Add cache to dataset. Move normalization out of augmentation.
* Trimesh.load can take up to 5 seconds to load, I didn't investigate
the reason, but just let it sample points and then save the to cache.
They're anyway will be shuffled in the augmentation.

* Putting the normalization in the augmentation step was a huge mistake,
since it's dramatically change input feature's scale and in validation time
the network produces crazy result. But funny, that it identify most of the
objects as a bookshelf.